Separate jest config, Add alias
