+++ package.json
@@ -2,7 +2,7 @@
   "name": "article-management-api",
   "version": "1.0.1",
   "description": "Article Management API",
-  "author": "",
+  "author": "Deepak Tomar",
   "private": true,
   "license": "UNLICENSED",
   "scripts": {
@@ -63,23 +63,6 @@
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
     "typescript": "^5.6.2"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "ts"
-    ],
-    "rootDir": "src",
-    "testRegex": ".*\\.spec\\.ts$",
-    "transform": {
-      "^.+\\.(t|j)s$": "ts-jest"
-    },
-    "collectCoverageFrom": [
-      "**/*.(t|j)s"
-    ],
-    "coverageDirectory": "../coverage",
-    "testEnvironment": "node"
   },
   "husky": {
     "hooks": {
