+++ src/modules/adapters/article.adapter.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common'
-import { DatabaseService } from './../database/database.service'
-import { TableName, ArticleEntry } from './../../types'
+import { DatabaseService } from 'src/modules/database/database.service'
+import { TableName, ArticleEntry } from 'src/types'
 
 @Injectable()
 export class ArticleAdapter extends DatabaseService<ArticleEntry> {
