+++ src/modules/adapters/author.adapter.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common'
-import { DatabaseService } from './../database/database.service'
-import { TableName, AuthorEntry } from './../../types'
+import { DatabaseService } from 'src/modules/database/database.service'
+import { TableName, AuthorEntry } from 'src/types'
 
 @Injectable()
 export class AuthorAdapter extends DatabaseService<AuthorEntry> {
