+++ src/modules/article/article.service.spec.ts
@@ -2,10 +2,10 @@ import { NotFoundException } from '@nestjs/common'
 import { Test, TestingModule } from '@nestjs/testing'
 import { CACHE_MANAGER } from '@nestjs/cache-manager'
 import { ArticleService } from './article.service'
-import { AuthorAdapter } from '../adapters/author.adapter'
-import { CommentAdapter } from '../adapters/comment.adapter'
-import { ArticleAdapter } from '../adapters/article.adapter'
-import { message } from './../../utils'
+import { AuthorAdapter } from 'src/modules/adapters/author.adapter'
+import { CommentAdapter } from 'src/modules/adapters/comment.adapter'
+import { ArticleAdapter } from 'src/modules/adapters/article.adapter'
+import { message } from 'src/utils'
 
 const mockCacheManager = {
   set: jest.fn(),
