+++ src/modules/authentication/authentication.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common'
 import { JwtModule } from '@nestjs/jwt'
-import { AdapterModule } from '../adapters/adapter.module'
-import { AuthenticationController } from '../authentication/authentication.controller'
-import { AuthenticationService } from '../authentication/authentication.service'
-import { SECRET_KEY, EXPIRATION_TIME } from './../../utils/constants'
+import { AdapterModule } from 'src/modules/adapters/adapter.module'
+import { AuthenticationController } from 'src/modules/authentication/authentication.controller'
+import { AuthenticationService } from 'src/modules/authentication/authentication.service'
+import { SECRET_KEY, EXPIRATION_TIME } from 'src/utils'
 
 @Module({
   imports: [
