+++ src/modules/author/author.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common'
-import { AdapterModule } from '../adapters/adapter.module'
+import { AdapterModule } from 'src/modules/adapters/adapter.module'
 import { AuthorController } from './author.controller'
 import { AuthorService } from './author.service'
