+++ src/modules/comment/comment.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common'
-import { AdapterModule } from '../adapters/adapter.module'
+import { AdapterModule } from 'src/modules/adapters/adapter.module'
 import { CommentController } from './comment.controller'
 import { CommentService } from './comment.service'
