+++ src/modules/comment/comment.service.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common'
-import { AuthorAdapter } from '../adapters/author.adapter'
-import { ArticleAdapter } from '../adapters/article.adapter'
-import { CommentAdapter } from '../adapters/comment.adapter'
-import { Comment } from './../../types'
-import { message } from './../../utils'
+import { AuthorAdapter } from 'src/modules/adapters/author.adapter'
+import { ArticleAdapter } from 'src/modules/adapters/article.adapter'
+import { CommentAdapter } from 'src/modules/adapters/comment.adapter'
+import { Comment } from 'src/types'
+import { message } from 'src/utils'
 
 @Injectable()
 export class CommentService {
